chore(frontend): move userOverview to components
